Refactor code responsible for book logic
