+++ bookLogic.rb
@@ -1,0 +1,29 @@
+require './Data'
+require './book'
+
+class BookLogic
+
+    include Data
+    
+    attr_accessor :books
+    def initialize
+        @books = Data.books
+    end
+
+    def list_all_books
+        puts 'Database is empty! Add a book.' if @books.empty?
+        @books.each { |book| puts "[Book] Title: #{book.title}, Author: #{book.author}" }
+    end
+
+    def create_book()
+        puts 'Create a new book'
+        print 'Enter title: '
+        title = gets.chomp
+        print 'Enter author: '
+        author = gets
+        book = Book.new(title, author)
+        @books.push(book)
+        puts "Book #{title} created successfully."
+      end
+
+end
